// src/enums/enums.ts
export enum RoutesEnum {
  Home = "/",
  Login = "/login",

  Dashboard = "/painel/dashboard",

  Produtos = "/painel/produtos",
  Cadastrar_Produto = "/painel/produtos/cadastrar",
  Editar_Produto = "/painel/produtos/:id",

  Vendas = "/painel/vendas",
  Editar_Venda = "/painel/vendas/:id",
  Cadastrar_Venda = "/painel/vendas/cadastrar",

  Compras = "/painel/compras",
  Editar_Compra = "/painel/compras/:id",
  Cadastrar_Compra = "/painel/compras/cadastrar",

  Relatorios = "/painel/relatorios",
  Register = "/painel/register",
  Configuracoes = "/painel/configuracoes",
}